Criação de service para cliente, alteração de get do tipo integer para long

+++ src/test/java/com/trabalhoFinal/trabalhoFinal/models/Cliente.java
@@ -54,7 +54,7 @@ public class Cliente implements Serializable {
     public Cliente() {
     }
 
-    public Integer getCodCliente() {
+    public Long getCodCliente() {
         return codCliente;
     }
